backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg2-binary>=2.9
django-cors-headers>=4.4
python-dotenv>=1.0
gunicorn>=22.0